Review - Moved Memberships section to its own component
